Major update to the script
Added a "Default IP Address" capability. Create ip.txt in the folder you run the script from. Put the IP address or range in the file. 
Added a switch for Linux vs Windows. sudo is now added to the script if Linux is chosen.
Several usability and visual enhancements.
Added an MS17-010 "WannaCry" script